Add utils.f90 and use stop_error()
The main problem is, that we need to test in our test suite that the library
works. If the library exits using:

    stop "some error message"

Then the main program will still exit with a zero exit code, so our test suite
passes... To fix this, one should use the "stop_error" subroutine:

    call stop_error("some error message")

That subroutine prints the message and calls "stop 1", which actually returns
a non-zero exit code and our test suite will fail.

Note that after this patch, our test will fail with a non-zero exit code and
Travis test will thus fail.